+++ apps/create.go
@@ -5,7 +5,6 @@ import (
 
 	"gopkg.in/errgo.v1"
 
-	"github.com/Scalingo/cli/appdetect"
 	"github.com/Scalingo/cli/config"
 	"github.com/Scalingo/cli/utils"
 	"github.com/Scalingo/go-scalingo/v4"
@@ -40,7 +39,7 @@ func Create(appName string, remote string, buildpack string) error {
 	}
 
 	fmt.Printf("App '%s' has been created\n", app.Name)
-	if _, ok := appdetect.DetectGit(); ok && appdetect.AddRemote(app.GitUrl, remote) == nil {
+	if _, ok := utils.DetectGit(); ok && utils.AddGitRemote(app.GitUrl, remote) == nil {
 		fmt.Printf("Git repository detected: remote %s added\n→ 'git push %s master' to deploy your app\n", remote, remote)
 	} else {
 		fmt.Printf("To deploy your application, run these commands in your GIT repository:\n→ git remote add %s %s\n→ git push %s master\n", remote, app.GitUrl, remote)