+++ cmd/destroy.go
@@ -3,9 +3,9 @@ package cmd
 import (
 	"github.com/urfave/cli"
 
-	"github.com/Scalingo/cli/appdetect"
 	"github.com/Scalingo/cli/apps"
 	"github.com/Scalingo/cli/cmd/autocomplete"
+	"github.com/Scalingo/cli/detect"
 )
 
 var (
@@ -26,7 +26,7 @@ var (
 				if len(c.Args()) != 0 {
 					currentApp = c.Args()[0]
 				} else {
-					currentApp = appdetect.CurrentApp(c)
+					currentApp = detect.CurrentApp(c)
 				}
 
 				err := apps.Destroy(currentApp, c.Bool("force"))