+++ cmd/one_off_stop.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/urfave/cli"
 
-	"github.com/Scalingo/cli/appdetect"
 	"github.com/Scalingo/cli/apps"
 	"github.com/Scalingo/cli/cmd/autocomplete"
+	"github.com/Scalingo/cli/detect"
 )
 
 var (
@@ -21,7 +21,7 @@ var (
 	  'scalingo --app my-app one-off-stop one-off-1234'
 	  'scalingo --app my-app one-off-stop 1234'`,
 		Action: func(c *cli.Context) {
-			currentApp := appdetect.CurrentApp(c)
+			currentApp := detect.CurrentApp(c)
 			if len(c.Args()) != 1 {
 				cli.ShowCommandHelp(c, "one-off-stop")
 				return