+++ cmd/stats.go
@@ -3,8 +3,8 @@ package cmd
 import (
 	"github.com/urfave/cli"
 
-	"github.com/Scalingo/cli/appdetect"
 	"github.com/Scalingo/cli/apps"
+	"github.com/Scalingo/cli/detect"
 )
 
 var (
@@ -20,7 +20,7 @@ var (
 	Example
 	  'scalingo --app my-app stats'`,
 		Action: func(c *cli.Context) {
-			currentApp := appdetect.CurrentApp(c)
+			currentApp := detect.CurrentApp(c)
 			if len(c.Args()) != 0 {
 				cli.ShowCommandHelp(c, "stats")
 			} else if err := apps.Stats(currentApp, c.Bool("stream")); err != nil {